# inflate: 10-15% faster decompression